# Conflicts:
#	fiit-mobile/app/src/main/java/ar/edu/itba/hci/fiit_mobile/data/network/api/ApiRoutineService.kt
#	fiit-mobile/app/src/main/java/ar/edu/itba/hci/fiit_mobile/data/network/api/ApiUserService.kt
#	fiit-mobile/app/src/main/java/ar/edu/itba/hci/fiit_mobile/ui/states/HomeUiState.kt

+++ fiit-mobile/app/src/main/java/ar/edu/itba/hci/fiit_mobile/data/network/model/executions/NetworkExecutionContent.kt
@@ -7,7 +7,7 @@ import com.google.gson.annotations.SerializedName
 data class NetworkExecutionContent (
 
     @SerializedName("id"          ) var id          : Int,
-    @SerializedName("date"        ) var date        : Int,
+    @SerializedName("date"        ) var date        : Long,
     @SerializedName("duration"    ) var duration    : Int,
     @SerializedName("wasModified" ) var wasModified : Boolean,
     @SerializedName("routine"     ) var routine     : NetworkRoutineContent? = null
